+++ src/IndexedDbStoreInfo.cs
@@ -5,4 +5,5 @@ internal class IndexedDbStoreInfo
     public string? DatabaseName { get; set; }
     public string? StoreName { get; set; }
     public int? Version { get; set; }
+    public string[]? StoreNames { get; set; }
 }